Move env related methods to separate source file
